Reorg (#11)
* Add DevOps + IT/Admin terms

* reorg

* update links

* add chapter 3-2

* Intro changes and suggestions

Some are typos/grammar, a couple are larger changes that I think improve flow. Feel free to use or not use whatever pieces make sense to you!

I didn't implement this, but it might be clearer if you lay out the two terms on line 51. It wasn't *immediately* clear that "IT/Admins" was the second term.

The Chapter List feels a little disjointed. Sometimes you use chapter titles, sometimes you don't; it's a table for Section 2, and lists for 1 and 3. I'd standardize things there as much as you can.

I assign the copyright of this contribution to Alex Gold.

(BTW, I think github warns me about the contributing thing if you put it in the `github` folder instead of the main folder. I'd PR that but I'm doing this edit in the web editor while I read.)

* Add DevOps + IT/Admin terms

* sec1 intro typos

I assign the copyright of this contribution to Alex Gold.

General review note: I find it somewhat confusing that the ### bullets don't match the chapter titles. Maybe something like:

### Use CI/CD ➡️ Code promotion and integration

* formatting changes

* Add DevOps + IT/Admin terms

* add chapter 3-2

* formatting changes

* add chapter 3-2

* Intro changes and suggestions

Some are typos/grammar, a couple are larger changes that I think improve flow. Feel free to use or not use whatever pieces make sense to you!

I didn't implement this, but it might be clearer if you lay out the two terms on line 51. It wasn't *immediately* clear that "IT/Admins" was the second term.

The Chapter List feels a little disjointed. Sometimes you use chapter titles, sometimes you don't; it's a table for Section 2, and lists for 1 and 3. I'd standardize things there as much as you can.

I assign the copyright of this contribution to Alex Gold.

(BTW, I think github warns me about the contributing thing if you put it in the `github` folder instead of the main folder. I'd PR that but I'm doing this edit in the web editor while I read.)

Co-authored-by: Jon Harmon <jonthegeek@gmail.com>