Chapter 5 tweaks and notes (#111)
* Chapter 5 tweaks and notes

Line 104: This is the second time you tried to point us away from your book. I think the mention at the start of the chapter is enough, don't feel like you have to keep doing that!

Line 110: Link doesn't go anywhere. Is that going to be an appendix? It's probably good to explicitly call out where you're sending us, since eventually this will be print and you can't rely on the link.

The docker example: I'm getting this in the browser when I run it:
```
{
error: "404 - Resource Not Found"
}
```

It seems like it built properly. Did I do something wrong, or is the image broken?

Line 159: Probably explicitly call out that we need to `exit` the docker terminal (to get back to our actual command line) before running the next docker command.

Oh hey, looking at the log told me to actually look at http://127.0.0.1:8000/__docs__/ to see what you were pointing at. http://localhost:8000/__docs__/ also works. I'm guessing maybe you lost the underscores in a copy/paste?

Line 163: Probably have us `docker kill palmer-plumber` again before moving on.

Line 171: Didn't work for me in my Windows Powershell. I get this error:

```
docker: invalid reference format.
See 'docker run --help'.
```

Removing the `$(pwd):` part allowed it to run, but so far I can't find where/if this actually mounted. I *think* you're saying I should see it in my normal file browser, but you might want to be more explicit about that in Line 175. I can see it in Volumes in Docker Desktop, but as far as I can tell that isn't corresponding to anywhere on my file system. I also tried changing `$(pwd):` to `.`, with similar "ran but no idea if it saved anything" results. I assume this is a Windows issue, but it's probably worth figuring out!

Line 214: Again the link doesn't go anywhere. New section coming?

Line 227: A concrete example would be helpful, even if it's fake.

Aha, now that I'm in the section on flags and grok the `:`, I got the volume stuff working using `${PWD}$:/project-out` (suggested [here](https://stackoverflow.com/questions/41485217/mount-current-directory-as-a-volume-in-docker-on-windows-10)). Supposedly the `${PWD}` version will work on Linux as well. If so, I recommend switching to that. If updating it works, be sure to update the references at lines 269 and 283 as well.

Line 366: I like the idea of an ongoing project, but it makes it hard to start in a particular chapter. We're seeing this issue since the projects didn't exist when we started; we can't continue a project we haven't started!

* Clean up docker's intro

Oops, I read this part before I started editing and forgot that this sentence threw me off.