Chapter 4 typos and notes (#110)
Line 49: It's unclear which context is "that context." I spelled it out and hope I wasn't confused 🙃

Line 107: Um, what's "trace"? Is that a synonym for "info" in this context? It looks like it is from context (and from what I know about usage), but it isn't obvious from the text.

Line 239: I'd like to see an example of this to solidify how and why to do it. Even if it's just screenshots or something rather than a full step-by-step/working app.

Line 249: Can you point us to a help doc that will get us there? I know links can be unstable, but maybe tell us (in words) where to look on the Grafana site (already linked) for details.

I'd like to see screenshots or... something... to tell us what we'd do with them, though. Or maybe the upcoming chapter will go into that more?

I love the new Comprehension Questions and Portfolio Exercise sections SOO much. I'm glad we built a spot in the club to go back and see what has changed, 'cuz we'll definitely have to go back and catch up with those!